api/requirements.txt
Flask==2.1.2
pymongo==3.12.1
python-dotenv==0.19.2
bcrypt==4.1.2
pyjwt==2.8.0
flask-jwt-extended==4.6.0
flask-cors==3.0.10
werkzeug>=2.0.0,<2.2.0